chore!: use u64 for NoiseLevel
Change from usize to u64 for MaxNoiseLevel and NoiseLevel

This is an API break as `new` and `get` handle/returns u64
instead of usize

This is also a potential serialization break depending on the
serializer used (bincode should be fine as it serializes usize as u64)

+++ tfhe/src/integer/server_key/radix_parallel/scalar_add.rs
@@ -555,7 +555,8 @@ impl ServerKey {
         let num_bits_in_block = packed_modulus.ilog2();
         // Just in case we compare with max noise level, but it should always be num_bits_in_blocks
         // with the parameters we provide
-        let grouping_size = (num_bits_in_block as usize).min(self.key.max_noise_level.get());
+        let grouping_size =
+            (num_bits_in_block as usize).min(self.key.max_noise_level.get() as usize);
 
         // In this, we store lookup tables to be used on each 'packing'.
         // These LUTs will generate an output that tells whether the packing
